+++ floship_llm/__init__.py
@@ -61,4 +61,4 @@ __all__ = [
     "lm_json_utils",
 ]
 
-__version__ = "1.4.1"
+__version__ = "1.4.2"